add move of TVL to hackint
We moved discussions of our monorepo-tooling and adjacent projects to hackint. Our old channel was one of the first ones taken over by Freenode staff because the move was mentioned in the topic (they kicked everyone out, but didn't actually seize control of the channel - still it motivated us to speed up the move).